Fixed issue with unserializing numeric keys
We need to mimic PHP's behaviour of $a[1] being equivalent to $a['1'].
Unfortunately, when you JSON encode an array it may convert it to an
object, and when you decode, you now have an object with numeric
properties. If you cast this directly to an array, you'll end up with
values that cannot be accessed directly.